Add .gitignore and remove .vs directory from tracking

+++ ConsoleApp1/Operacje.cs
@@ -4,10 +4,14 @@ namespace Kalkulator
 {
     public class Operacje
     {
-        public static Func<int, int, int> dodaj = (a, b) => a + b;
-        public static Func<int, int, int> odejmij = (a, b) => a - b;
-        public static Func<int, int, int> pomnoz = (a, b) => a * b;
-        public static Func<int, int, int> podziel = (a, b) => a / b;
-        public static Func<int, int, int> procent = (a, b) => (a * b) / 100 ;
+        public static Func<double, double, double> dodaj = (a, b) => a + b;
+
+        public static Func<double, double, double> odejmij = (a, b) => a - b;
+
+        public static Func<double, double, double> pomnoz = (a, b) => a * b;
+
+        public static Func<double, double, double> podziel = (a, b) => a / b;
+
+        public static Func<double, double, double> procent = (a, b) => (a * b) / 100 ;
     }
 }
